Fix bugs, output looks ok now
Removed some format check when reading in files. istream_iterator
doesn't become the so-called end-of-stream iterator (which is returned
by the default constructor) on eof, only after fail. That would make
range based for loop less elegant and I'm lazy, so I just don't care if
fail or eof happens.